requirements.txt
requests==2.31.0
selenium==4.15.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
schedule==1.2.0
cohere==4.37
linkedin-api==2.0.0 